requirements.txt
Flask==3.0.3
numpy==1.24.4
pandas==1.5.3
plotly==5.24.1
dash==2.18.2
watchdog==6.0.0
flask-apscheduler
flask-wtf
sqlalchemy
